% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_fit.R
\name{pg_ensemble}
\alias{pg_ensemble}
\title{Fit spatial Bayesian ensemble}
\usage{
pg_ensemble(dens, X, space_id, coords, n_iter)
}
\arguments{
\item{dens}{Matrix of model densities (N x M)}

\item{X}{Matrix of covariates (N x P)}

\item{space_id}{Spatial location ID vector (N)}

\item{coords}{Matrix of x y coordinates, with colnames(coords) == c("x", "y"), (N, 2)}

\item{n_iter}{Number of iterations used in MCMC}
}
\value{
A list containing MCMC output
}
\description{
This function fits a spatial Bayesian ensemble using model densities and (optional) ensemble weight covariates
}
\examples{
# pg_ensemble()


}
