% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_pred.R
\name{pg_pred}
\alias{pg_pred}
\title{Fit the Geostatistical Regression Model (GRM)}
\usage{
pg_pred(pg_fit, X, space_id, coords)
}
\arguments{
\item{pg_fit}{Ensemble fit object created with pg_ensemble()}

\item{X}{Matrix of covariates (N_pred x P)}

\item{space_id}{spatial location ID vector (N_pred)}

\item{coords}{Matrix of prediction x y coordinates, with colnames(coords) == c("x", "y"), (N_pred, 2)}
}
\value{
A list containing MCMC output
}
\description{
This function fits a spatial Bayesian ensemble using model densities and (optional) ensemble weight covariates
}
\examples{
# pg_pred()


}
