% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{la_grid}
\alias{la_grid}
\title{Community Multiscale Air Quality (CMAQ) Chemical Transport Model (CMT) Data For Atlanta Metro Area, June 2004}
\format{
\subsection{\code{cmaq_full}}{

A data frame with 72,000 rows and 15 columns:
\describe{
\item{time_id}{Week identifier}
\item{space_id}{Location identifier}
\item{spacetime_id}{Year identifier}
\item{aod}{Aerosol optical depth value}
\item{elevation}{Elevation spatial covariate (US Geological Survey)}
\item{forestcover}{Percentage of forest cover spatial covariate (2001 National Land Cover database)}
\item{hwy_length}{Sum of major roadway length spatial covariate (2001 National Land Cover database)}
\item{lim_hwy_length}{Sum of major roadway length spatial covariate (2001 National Land Cover database)}
\item{lim_hwy_length}{Sum of local roadway lengths spatial covariate (2001 National Land Cover database)}
\item{point_emi_any}{Indicator of PM2.5 primary emission point source spatial covariate (2002 USEPA National Emissions Inventory)}
\item{tmp}{Temperature spatio-temporal covariate (North American Land Data Assimilation Systems)}
\item{wind}{Wind spatio-temporal covariate (North American Land Data Assimilation Systems)}
\item{ctm}{Chemical transport model value}
\item{date}{Date of measurement}
\item{x}{Location x-coordinate}
\item{y}{Location y-coordinate}
\item{cmaqgrm_model_estimate}{PM2.5 estimate from CMAQ-GRM model.}
\item{cmaqgrm_model_sd}{PM2.5 standard deviation from CMAQ-GRM model.}
\item{modisgrm_model_estimate}{PM2.5 estimate from MODIS-GRM model.}
\item{modisgrm_model_sd}{PM2.5 standard deviation from MODIS-GRM model.}
\item{bart_model_estimate}{PM2.5 estimate from BART model.}
\item{bart_model_sd}{PM2.5 standard deviation from BART model.}
...
}
}
}
\source{
\url{https://www.nature.com/articles/jes201390}
}
\usage{
la_grid
}
\description{
Daily CMAQ CMT output, MODIS AOD readings, covariates of interest, and CMAQ-GRM, MODIS-GRM, and BART model prediction outputs, in Atlanta Metro Area for June 2004.
Data included for every 12km x 12km grid square in study area.
Each row of the data is a unique location/day.
}
\keyword{datasets}
